app/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-multipart==0.0.12  # Для обробки form-data з фото
pillow==11.1.0           # Для валідації зображень